✨ first version that sort of works
