+++ adminSiteClient/admin.entry.ts
@@ -1,4 +1,5 @@
 import "./admin.scss"
+import "../site/owid.scss"
 import "@ourworldindata/grapher/src/core/grapher.scss"
 import "../explorerAdminClient/ExplorerCreatePage.scss"
 import "handsontable/dist/handsontable.full.css"
